+++ packages/astro/src/vite-plugin-astro-server/devPipeline.ts
@@ -4,11 +4,11 @@ import type {
 	RuntimeMode,
 	SSRLoadedRenderer,
 	SSRManifest,
-} from '../@types/astro';
-import type { Logger } from '../core/logger/core';
-import type { ModuleLoader } from '../core/module-loader';
+} from '../@types/astro.js';
+import type { Logger } from '../core/logger/core.js';
+import type { ModuleLoader } from '../core/module-loader/index.js';
 import { Pipeline } from '../core/pipeline.js';
-import type { Environment } from '../core/render';
+import type { Environment } from '../core/render/index.js';
 import { createEnvironment, loadRenderer } from '../core/render/index.js';
 import { RouteCache } from '../core/render/route-cache.js';
 import { isServerLikeOutput } from '../prerender/utils.js';