+++ packages/astro/test/core-image.test.js
@@ -3,10 +3,10 @@ import * as cheerio from 'cheerio';
 import { basename } from 'node:path';
 import { Writable } from 'node:stream';
 import { removeDir } from '../dist/core/fs/index.js';
+import { Logger } from '../dist/core/logger/core.js';
 import testAdapter from './test-adapter.js';
 import { testImageService } from './test-image-service.js';
 import { loadFixture } from './test-utils.js';
-import { Logger } from '../dist/core/logger/core.js';
 
 describe('astro:image', () => {
 	/** @type {import('./test-utils').Fixture} */
@@ -158,6 +158,37 @@ describe('astro:image', () => {
 				let src = $img.attr('src');
 				res = await fixture.fetch(src);
 				expect(res.status).to.equal(200);
+			});
+
+			it('supports uppercased imports', async () => {
+				let res = await fixture.fetch('/uppercase');
+				let html = await res.text();
+				$ = cheerio.load(html);
+
+				let $img = $('img');
+				expect($img).to.have.a.lengthOf(1);
+
+				let src = $img.attr('src');
+				let loading = $img.attr('loading');
+				res = await fixture.fetch(src);
+				expect(res.status).to.equal(200);
+				expect(loading).to.not.be.undefined;
+			});
+
+			it('supports avif', async () => {
+				let res = await fixture.fetch('/avif');
+				let html = await res.text();
+				$ = cheerio.load(html);
+
+				console.log(html);
+
+				let $img = $('img');
+				expect($img).to.have.a.lengthOf(1);
+
+				let src = $img.attr('src');
+				res = await fixture.fetch(src);
+				expect(res.status).to.equal(200);
+				expect(res.headers.get('content-type')).to.equal('image/avif');
 			});
 		});
 
@@ -429,6 +460,47 @@ describe('astro:image', () => {
 
 				const $ = cheerio.load(html);
 				expect($('#local img').attr('data-service-config')).to.equal('bar');
+			});
+		});
+
+		describe('custom endpoint', async () => {
+			/** @type {import('./test-utils').DevServer} */
+			let customEndpointDevServer;
+
+			/** @type {import('./test-utils.js').Fixture} */
+			let customEndpointFixture;
+
+			before(async () => {
+				customEndpointFixture = await loadFixture({
+					root: './fixtures/core-image/',
+					image: {
+						endpoint: './src/custom-endpoint.ts',
+						service: testImageService({ foo: 'bar' }),
+						domains: ['avatars.githubusercontent.com'],
+					},
+				});
+
+				customEndpointDevServer = await customEndpointFixture.startDevServer({
+					server: { port: 4324 },
+				});
+			});
+
+			it('custom endpoint works', async () => {
+				const response = await customEndpointFixture.fetch('/');
+				const html = await response.text();
+
+				const $ = cheerio.load(html);
+				const src = $('#local img').attr('src');
+
+				let res = await customEndpointFixture.fetch(src);
+				expect(res.status).to.equal(200);
+				expect(await res.text()).to.equal(
+					"You fool! I'm not a image endpoint at all, I just return this!"
+				);
+			});
+
+			after(async () => {
+				await customEndpointDevServer.stop();
 			});
 		});
 	});