+++ packages/create-astro/src/actions/next-steps.ts
@@ -1,20 +1,23 @@
 import path from 'node:path';
-import type { Context } from './context';
+import type { Context } from './context.js';
 
 import { nextSteps, say } from '../messages.js';
 
-export async function next(ctx: Pick<Context, 'cwd' | 'pkgManager' | 'skipHouston'>) {
+export async function next(ctx: Pick<Context, 'hat' | 'cwd' | 'packageManager' | 'skipHouston'>) {
 	let projectDir = path.relative(process.cwd(), ctx.cwd);
-	const devCmd =
-		ctx.pkgManager === 'npm'
-			? 'npm run dev'
-			: ctx.pkgManager === 'bun'
-			? 'bun run dev'
-			: `${ctx.pkgManager} dev`;
+
+	const commandMap: { [key: string]: string } = {
+		npm: 'npm run dev',
+		bun: 'bun run dev',
+		yarn: 'yarn dev',
+		pnpm: 'pnpm dev',
+	};
+
+	const devCmd = commandMap[ctx.packageManager as keyof typeof commandMap] || 'npm run dev';
 	await nextSteps({ projectDir, devCmd });
 
 	if (!ctx.skipHouston) {
-		await say(['Good luck out there, astronaut! 🚀']);
+		await say(['Good luck out there, astronaut! 🚀'], { hat: ctx.hat });
 	}
 	return;
 }