+++ packages/markdown/remark/test/test-utils.js
@@ -1,3 +1,0 @@
-export const mockRenderMarkdownParams = {
-	contentDir: new URL('file:///src/content/'),
-};